Refactor to allow generation of resources from multiple images.
This is done to allow generated Ingress resources to be flattened so
that one Ingress can be generated for a number of Docker images.

Along with this change, it's now necessary for the operator to specify
which resource types they would like to generate, as this fits better
with the workflow where multiple images are being provided.