Expose TargetRuntimeArgs, with a package private constructor
TargetRuntimeArgs is the type of targetRuntimeArgs, which is a public
property of GenerateDeploymentArgs, which in turn is public. So
TargetRuntimeArgs itself should be a public type.

To restrict it being extended by anyone (which is, I believe, why it was
package private) we can give it a package private constructor, by also
making it a class rather than a trait.