Fix the build for Ubuntu 18.04
At this time, Bintray cannot understand the new format generated
by default by dpkg 1.19 or later. There is a new flag that is
available in "dpkg-deb", but not in "dpkg"; also, the flag is
not understood by older versions. This patch detects the dpkg
version, and enables the compatibility flag when necessary.

Fixes #144