Authentication fixes: Basic & EncodedBasic
The current authentication code relies on a
Bearer Token to be generated at some point.
Some docker repositories, instead, rely on
basic authentication, notably AWS ECR, or on
basic tokens (not bearer tokens).

This commit allows plain basic authentication
to succeed.

In a similar vein, the code was also not using
effectively the encoded credentials saved in
`~/.docker/config.json` for those repositories
that have already received a "docker login";
it will now use that information if available
to access the repository.