Generate a dummy LoadBalancer service
During Akka Cluster Boostrap using DNS, the self reference using IP address must
match up with the replicas we find via DNS.
The SRV record returned by the headless service is in the form of `auto-generated-name.my-svc.my-namespace.svc.cluster.local`.
https://kubernetes.io/docs/concepts/services-networking/dns-pod-service/#srv-records

Under some circumstances DNS also returns A record for the pods in the form of `pod-ip-address.my-namespace.pod.cluster.local`,
which we require to correctly figure out the lowest IP address (and to identify self node).
For reasons yet to be confirmed, service type `ClusterIP` (default) doesn't always work, but `LoadBalancer` works.
This means that the service will expose the service externally "using a cloud provider’s load balancer."
https://kubernetes.io/docs/concepts/services-networking/service/#publishing-services-service-types

To workaround this, this creates a dummy service that exposes port 70 for Gopher.